BIGINT des ID pour etre coherent avec l'api USER

+++ src/main/resources/db/migration/V1_0__image.sql
@@ -1,8 +1,8 @@
 CREATE TABLE image (
-	id INTEGER AUTO_INCREMENT,
+	id BIGINT AUTO_INCREMENT,
 	file BLOB,
 	titre VARCHAR(255),
-	id_utilisateur INTEGER,
+	id_utilisateur BIGINT,
 	date DATE,
 	PRIMARY KEY(id)
 )
